Odděl ingestion API test od PostgreSQL acceptance
